[zero] refactor gemini folder structure
