+++ colossalai/zero/gemini/chunk/utils.py
@@ -5,9 +5,10 @@ import torch
 import torch.distributed as dist
 import torch.nn as nn
 
-from colossalai.gemini.chunk import ChunkManager
-from colossalai.gemini.chunk.search_utils import search_chunk_configuration
 from colossalai.utils import is_ddp_ignored
+
+from .manager import ChunkManager
+from .search_utils import search_chunk_configuration
 
 
 def safe_div(a, b):
