+++ colossalai/zero/gemini/placement_policy.py
@@ -5,10 +5,11 @@ from typing import Dict, List, Optional, Tuple, Type
 
 import torch
 
-from colossalai.gemini.chunk import Chunk, ChunkManager
-from colossalai.gemini.memory_tracer import ChunkMemStatsCollector
 from colossalai.utils import get_current_device
 from colossalai.utils.memory import colo_device_memory_capacity
+
+from .chunk import Chunk, ChunkManager
+from .memory_tracer import ChunkMemStatsCollector
 
 
 class PlacementPolicy(ABC):
